+++ src/SpaceLiDAR.jl
@@ -24,9 +24,10 @@ include("ICESat/GLAH14.jl")
 include("geoid.jl")
 include("table.jl")
 
-export find, download!, netrc!, instantiate
-export xyz, lines, points, in_bbox, bounds
-export test, granule_from_file, granules_from_folder, write_granule_urls!
+export find, download!, netrc!, instantiate, info, angle, angle!, shift
+export lines, points, in_bbox, bounds, classify, isvalid, rm, to_egm2008!
+export ICESat_Granule, ICESat2_Granule, GEDI_Granule, convert
+export granule_from_file, granules_from_folder, write_granule_urls!
 
 precompile(find, (Symbol, String, NamedTuple, String))
 precompile(find, (Symbol, String))